Add Netezza support
In the replacementPatterns.csv file, copy the sql server to postgresql
patterns, replace postgresql with netezza as the target dbms, and append
these new lines to the end of the file. These patterns are expected to
mostly work based on Netezza's claim to support postgresql syntax, but
need to be tested. Amend the documentation to add Netezza as a target
and mention that the Netezza support is in beta pending testing.

Signed-off-by: Aaron Browne <aaron0browne@gmail.com>